# api/rest/gateway_rest.pb.go
// Code generated by protoc-gen-go-rest. DO NOT EDIT.
// versions:
// - protoc-gen-go-rest v0.0.0-alpha.0
// - protoc             v7.34.0
// source: gateway.proto

package rest

import (
	context "context"
	json "encoding/json"
	runtime "github.com/merzzzl/proto-rest-api/runtime"
	emptypb "google.golang.org/protobuf/types/known/emptypb"
	io "io"
	http "net/http"
)

// MeshServiceWebServer is the server API for MeshService service.
// All implementations must embed UnimplementedMeshServiceWebServer for forward compatibility.
type MeshServiceWebServer interface {
	// ListNodes returns a compact summary of every node the radio can see.
	ListNodes(context.Context, *ListNodesRequest) (*ListNodesResponse, error)
	// GetNode returns all stored information for one node by its number.
	GetNode(context.Context, *GetNodeRequest) (*Node, error)
	// ListNodeMessages returns direct messages received from one node.
	ListNodeMessages(context.Context, *ListMessagesRequest) (*ListMessagesResponse, error)
	// SendNodeMessage transmits a text message to one node.
	SendNodeMessage(context.Context, *SendMessageRequest) (*emptypb.Empty, error)
	// ListChannels returns every configured channel on the radio.
	ListChannels(context.Context, *ListChannelsRequest) (*ListChannelsResponse, error)
	// ListChannelMessages returns messages received on one channel.
	ListChannelMessages(context.Context, *ListChannelMessagesRequest) (*ListMessagesResponse, error)
	mustEmbedUnimplementedMeshServiceWebServer()
}

// UnimplementedMeshServiceWebServer must be embedded to have forward compatible implementations.
type UnimplementedMeshServiceWebServer struct{}

func (UnimplementedMeshServiceWebServer) ListNodes(context.Context, *ListNodesRequest) (*ListNodesResponse, error) {
	return nil, runtime.Errorf(http.StatusNotImplemented, "method not implemented")
}

func (UnimplementedMeshServiceWebServer) GetNode(context.Context, *GetNodeRequest) (*Node, error) {
	return nil, runtime.Errorf(http.StatusNotImplemented, "method not implemented")
}

func (UnimplementedMeshServiceWebServer) ListNodeMessages(context.Context, *ListMessagesRequest) (*ListMessagesResponse, error) {
	return nil, runtime.Errorf(http.StatusNotImplemented, "method not implemented")
}

func (UnimplementedMeshServiceWebServer) SendNodeMessage(context.Context, *SendMessageRequest) (*emptypb.Empty, error) {
	return nil, runtime.Errorf(http.StatusNotImplemented, "method not implemented")
}

func (UnimplementedMeshServiceWebServer) ListChannels(context.Context, *ListChannelsRequest) (*ListChannelsResponse, error) {
	return nil, runtime.Errorf(http.StatusNotImplemented, "method not implemented")
}

func (UnimplementedMeshServiceWebServer) ListChannelMessages(context.Context, *ListChannelMessagesRequest) (*ListMessagesResponse, error) {
	return nil, runtime.Errorf(http.StatusNotImplemented, "method not implemented")
}

func (UnimplementedMeshServiceWebServer) mustEmbedUnimplementedMeshServiceWebServer() {}

// CallbackServiceWebServer is the server API for CallbackService service.
// All implementations must embed UnimplementedCallbackServiceWebServer for forward compatibility.
type CallbackServiceWebServer interface {
	// RegisterCallback subscribes a webhook URL to mesh events.
	RegisterCallback(context.Context, *RegisterCallbackRequest) (*emptypb.Empty, error)
	// ListCallbacks returns every registered webhook subscriptions.
	ListCallbacks(context.Context, *ListCallbacksRequest) (*ListCallbacksResponse, error)
	// DeleteCallback unsubscribes a webhook URL.
	DeleteCallback(context.Context, *DeleteCallbackRequest) (*emptypb.Empty, error)
	mustEmbedUnimplementedCallbackServiceWebServer()
}

// UnimplementedCallbackServiceWebServer must be embedded to have forward compatible implementations.
type UnimplementedCallbackServiceWebServer struct{}

func (UnimplementedCallbackServiceWebServer) RegisterCallback(context.Context, *RegisterCallbackRequest) (*emptypb.Empty, error) {
	return nil, runtime.Errorf(http.StatusNotImplemented, "method not implemented")
}

func (UnimplementedCallbackServiceWebServer) ListCallbacks(context.Context, *ListCallbacksRequest) (*ListCallbacksResponse, error) {
	return nil, runtime.Errorf(http.StatusNotImplemented, "method not implemented")
}

func (UnimplementedCallbackServiceWebServer) DeleteCallback(context.Context, *DeleteCallbackRequest) (*emptypb.Empty, error) {
	return nil, runtime.Errorf(http.StatusNotImplemented, "method not implemented")
}

func (UnimplementedCallbackServiceWebServer) mustEmbedUnimplementedCallbackServiceWebServer() {}

// RegisterMeshServiceHandler registers the http handlers for service MeshService to "mux".
func RegisterMeshServiceHandler(router *runtime.Router, server MeshServiceWebServer, interceptors ...runtime.Interceptor) {
	if router == nil {
		return
	}

	router.Handle("GET", "/api/v1/nodes/:num", func(w http.ResponseWriter, r *http.Request, p runtime.Params) {
		handlerMeshServiceWebServerGetNode(server, w, r, p, interceptors)
	})

	router.Handle("GET", "/api/v1/channels/:index/messages", func(w http.ResponseWriter, r *http.Request, p runtime.Params) {
		handlerMeshServiceWebServerListChannelMessages(server, w, r, p, interceptors)
	})

	router.Handle("GET", "/api/v1/channels", func(w http.ResponseWriter, r *http.Request, p runtime.Params) {
		handlerMeshServiceWebServerListChannels(server, w, r, p, interceptors)
	})

	router.Handle("GET", "/api/v1/nodes/:num/messages", func(w http.ResponseWriter, r *http.Request, p runtime.Params) {
		handlerMeshServiceWebServerListNodeMessages(server, w, r, p, interceptors)
	})

	router.Handle("GET", "/api/v1/nodes", func(w http.ResponseWriter, r *http.Request, p runtime.Params) {
		handlerMeshServiceWebServerListNodes(server, w, r, p, interceptors)
	})

	router.Handle("POST", "/api/v1/nodes/:num/messages", func(w http.ResponseWriter, r *http.Request, p runtime.Params) {
		handlerMeshServiceWebServerSendNodeMessage(server, w, r, p, interceptors)
	})
}

// RegisterCallbackServiceHandler registers the http handlers for service CallbackService to "mux".
func RegisterCallbackServiceHandler(router *runtime.Router, server CallbackServiceWebServer, interceptors ...runtime.Interceptor) {
	if router == nil {
		return
	}

	router.Handle("DELETE", "/api/v1/callback", func(w http.ResponseWriter, r *http.Request, p runtime.Params) {
		handlerCallbackServiceWebServerDeleteCallback(server, w, r, p, interceptors)
	})

	router.Handle("GET", "/api/v1/callback", func(w http.ResponseWriter, r *http.Request, p runtime.Params) {
		handlerCallbackServiceWebServerListCallbacks(server, w, r, p, interceptors)
	})

	router.Handle("POST", "/api/v1/callback", func(w http.ResponseWriter, r *http.Request, p runtime.Params) {
		handlerCallbackServiceWebServerRegisterCallback(server, w, r, p, interceptors)
	})
}

func handlerMeshServiceWebServerListNodes(server MeshServiceWebServer, w http.ResponseWriter, r *http.Request, _ runtime.Params, il []runtime.Interceptor) {
	ctx, cancel := context.WithCancel(r.Context())
	defer cancel()

	w.Header().Set("Content-Type", "application/json")

	ctx, err := runtime.ApplyInterceptors(ctx, r, il...)
	if err != nil {
		errstatus := runtime.GetHTTPStatusFromError(err)

		w.WriteHeader(errstatus.Code())
		if _, err := w.Write([]byte(errstatus.Message())); err != nil {
			w.WriteHeader(http.StatusInternalServerError)
		}

		return
	}

	ctx = runtime.ContextWithHeaders(ctx, r.Header)

	var protoReq ListNodesRequest

	msg, err := server.ListNodes(ctx, &protoReq)
	if err != nil {
		errstatus := runtime.GetHTTPStatusFromError(err)

		w.WriteHeader(errstatus.Code())
		if _, err := w.Write([]byte(errstatus.Message())); err != nil {
			w.WriteHeader(http.StatusInternalServerError)
		}

		return
	}

	raw, err := json.Marshal(msg.GetNodes())
	if err != nil {
		w.WriteHeader(http.StatusInternalServerError)
		if _, err := w.Write([]byte(err.Error())); err != nil {
			w.WriteHeader(http.StatusInternalServerError)
		}

		return
	}

	w.WriteHeader(200)
	if _, err := w.Write(raw); err != nil {
		w.WriteHeader(http.StatusInternalServerError)
	}

	return

}

func (x *NodeSummary) UnmarshalJSON(data []byte) error {
	return runtime.ProtoUnmarshal(data, x)
}

func (x *NodeSummary) MarshalJSON() ([]byte, error) {
	return runtime.ProtoMarshal(x)
}

func handlerMeshServiceWebServerGetNode(server MeshServiceWebServer, w http.ResponseWriter, r *http.Request, p runtime.Params, il []runtime.Interceptor) {
	ctx, cancel := context.WithCancel(r.Context())
	defer cancel()

	w.Header().Set("Content-Type", "application/json")

	ctx, err := runtime.ApplyInterceptors(ctx, r, il...)
	if err != nil {
		errstatus := runtime.GetHTTPStatusFromError(err)

		w.WriteHeader(errstatus.Code())
		if _, err := w.Write([]byte(errstatus.Message())); err != nil {
			w.WriteHeader(http.StatusInternalServerError)
		}

		return
	}

	ctx = runtime.ContextWithHeaders(ctx, r.Header)

	var protoReq GetNodeRequest

	if v, err := runtime.ParseUint32(p.ByName("num")); err != nil {
		w.WriteHeader(http.StatusBadRequest)

		if _, err := w.Write([]byte(err.Error())); err != nil {
			w.WriteHeader(http.StatusInternalServerError)
		}

		return
	} else {
		protoReq.Num = v
	}

	msg, err := server.GetNode(ctx, &protoReq)
	if err != nil {
		errstatus := runtime.GetHTTPStatusFromError(err)

		w.WriteHeader(errstatus.Code())
		if _, err := w.Write([]byte(errstatus.Message())); err != nil {
			w.WriteHeader(http.StatusInternalServerError)
		}

		return
	}

	raw, err := runtime.ProtoMarshal(msg)
	if err != nil {
		w.WriteHeader(http.StatusInternalServerError)
		if _, err := w.Write([]byte(err.Error())); err != nil {
			w.WriteHeader(http.StatusInternalServerError)
		}

		return
	}

	w.WriteHeader(200)
	if _, err := w.Write(raw); err != nil {
		w.WriteHeader(http.StatusInternalServerError)
	}

	return

}

func handlerMeshServiceWebServerListNodeMessages(server MeshServiceWebServer, w http.ResponseWriter, r *http.Request, p runtime.Params, il []runtime.Interceptor) {
	ctx, cancel := context.WithCancel(r.Context())
	defer cancel()

	w.Header().Set("Content-Type", "application/json")

	ctx, err := runtime.ApplyInterceptors(ctx, r, il...)
	if err != nil {
		errstatus := runtime.GetHTTPStatusFromError(err)

		w.WriteHeader(errstatus.Code())
		if _, err := w.Write([]byte(errstatus.Message())); err != nil {
			w.WriteHeader(http.StatusInternalServerError)
		}

		return
	}

	ctx = runtime.ContextWithHeaders(ctx, r.Header)

	var protoReq ListMessagesRequest

	if v, err := runtime.ParseUint32(p.ByName("num")); err != nil {
		w.WriteHeader(http.StatusBadRequest)

		if _, err := w.Write([]byte(err.Error())); err != nil {
			w.WriteHeader(http.StatusInternalServerError)
		}

		return
	} else {
		protoReq.Num = v
	}

	msg, err := server.ListNodeMessages(ctx, &protoReq)
	if err != nil {
		errstatus := runtime.GetHTTPStatusFromError(err)

		w.WriteHeader(errstatus.Code())
		if _, err := w.Write([]byte(errstatus.Message())); err != nil {
			w.WriteHeader(http.StatusInternalServerError)
		}

		return
	}

	raw, err := json.Marshal(msg.GetMessages())
	if err != nil {
		w.WriteHeader(http.StatusInternalServerError)
		if _, err := w.Write([]byte(err.Error())); err != nil {
			w.WriteHeader(http.StatusInternalServerError)
		}

		return
	}

	w.WriteHeader(200)
	if _, err := w.Write(raw); err != nil {
		w.WriteHeader(http.StatusInternalServerError)
	}

	return

}

func (x *Message) UnmarshalJSON(data []byte) error {
	return runtime.ProtoUnmarshal(data, x)
}

func (x *Message) MarshalJSON() ([]byte, error) {
	return runtime.ProtoMarshal(x)
}

func handlerMeshServiceWebServerSendNodeMessage(server MeshServiceWebServer, w http.ResponseWriter, r *http.Request, p runtime.Params, il []runtime.Interceptor) {
	ctx, cancel := context.WithCancel(r.Context())
	defer cancel()

	w.Header().Set("Content-Type", "application/json")

	ctx, err := runtime.ApplyInterceptors(ctx, r, il...)
	if err != nil {
		errstatus := runtime.GetHTTPStatusFromError(err)

		w.WriteHeader(errstatus.Code())
		if _, err := w.Write([]byte(errstatus.Message())); err != nil {
			w.WriteHeader(http.StatusInternalServerError)
		}

		return
	}

	ctx = runtime.ContextWithHeaders(ctx, r.Header)

	var protoReq SendMessageRequest

	defer r.Body.Close()

	data, err := io.ReadAll(r.Body)
	if err != nil {
		w.WriteHeader(http.StatusBadRequest)
		if _, err := w.Write([]byte(err.Error())); err != nil {
			w.WriteHeader(http.StatusInternalServerError)
		}

		return
	}

	if len(data) != 0 {
		fm, err := runtime.GetFieldMaskJS(data)
		if err != nil {
			w.WriteHeader(http.StatusBadRequest)
			if _, err := w.Write([]byte(err.Error())); err != nil {
				w.WriteHeader(http.StatusInternalServerError)
			}

			return
		}

		ctx = runtime.ContextWithFieldMask(ctx, fm)
		err = runtime.ProtoUnmarshal(data, &protoReq)
		if err != nil {
			w.WriteHeader(http.StatusBadRequest)
			if _, err := w.Write([]byte(err.Error())); err != nil {
				w.WriteHeader(http.StatusInternalServerError)
			}

			return
		}
	}

	if v, err := runtime.ParseUint32(p.ByName("num")); err != nil {
		w.WriteHeader(http.StatusBadRequest)

		if _, err := w.Write([]byte(err.Error())); err != nil {
			w.WriteHeader(http.StatusInternalServerError)
		}

		return
	} else {
		protoReq.Num = v
	}

	_, err = server.SendNodeMessage(ctx, &protoReq)
	if err != nil {
		errstatus := runtime.GetHTTPStatusFromError(err)

		w.WriteHeader(errstatus.Code())
		if _, err := w.Write([]byte(errstatus.Message())); err != nil {
			w.WriteHeader(http.StatusInternalServerError)
		}

		return
	}

	w.WriteHeader(202)
	if _, err := w.Write(nil); err != nil {
		w.WriteHeader(http.StatusInternalServerError)
	}

	return

}

func handlerMeshServiceWebServerListChannels(server MeshServiceWebServer, w http.ResponseWriter, r *http.Request, _ runtime.Params, il []runtime.Interceptor) {
	ctx, cancel := context.WithCancel(r.Context())
	defer cancel()

	w.Header().Set("Content-Type", "application/json")

	ctx, err := runtime.ApplyInterceptors(ctx, r, il...)
	if err != nil {
		errstatus := runtime.GetHTTPStatusFromError(err)

		w.WriteHeader(errstatus.Code())
		if _, err := w.Write([]byte(errstatus.Message())); err != nil {
			w.WriteHeader(http.StatusInternalServerError)
		}

		return
	}

	ctx = runtime.ContextWithHeaders(ctx, r.Header)

	var protoReq ListChannelsRequest

	msg, err := server.ListChannels(ctx, &protoReq)
	if err != nil {
		errstatus := runtime.GetHTTPStatusFromError(err)

		w.WriteHeader(errstatus.Code())
		if _, err := w.Write([]byte(errstatus.Message())); err != nil {
			w.WriteHeader(http.StatusInternalServerError)
		}

		return
	}

	raw, err := json.Marshal(msg.GetChannels())
	if err != nil {
		w.WriteHeader(http.StatusInternalServerError)
		if _, err := w.Write([]byte(err.Error())); err != nil {
			w.WriteHeader(http.StatusInternalServerError)
		}

		return
	}

	w.WriteHeader(200)
	if _, err := w.Write(raw); err != nil {
		w.WriteHeader(http.StatusInternalServerError)
	}

	return

}

func (x *Channel) UnmarshalJSON(data []byte) error {
	return runtime.ProtoUnmarshal(data, x)
}

func (x *Channel) MarshalJSON() ([]byte, error) {
	return runtime.ProtoMarshal(x)
}

func handlerMeshServiceWebServerListChannelMessages(server MeshServiceWebServer, w http.ResponseWriter, r *http.Request, p runtime.Params, il []runtime.Interceptor) {
	ctx, cancel := context.WithCancel(r.Context())
	defer cancel()

	w.Header().Set("Content-Type", "application/json")

	ctx, err := runtime.ApplyInterceptors(ctx, r, il...)
	if err != nil {
		errstatus := runtime.GetHTTPStatusFromError(err)

		w.WriteHeader(errstatus.Code())
		if _, err := w.Write([]byte(errstatus.Message())); err != nil {
			w.WriteHeader(http.StatusInternalServerError)
		}

		return
	}

	ctx = runtime.ContextWithHeaders(ctx, r.Header)

	var protoReq ListChannelMessagesRequest

	if v, err := runtime.ParseUint32(p.ByName("index")); err != nil {
		w.WriteHeader(http.StatusBadRequest)

		if _, err := w.Write([]byte(err.Error())); err != nil {
			w.WriteHeader(http.StatusInternalServerError)
		}

		return
	} else {
		protoReq.Index = v
	}

	msg, err := server.ListChannelMessages(ctx, &protoReq)
	if err != nil {
		errstatus := runtime.GetHTTPStatusFromError(err)

		w.WriteHeader(errstatus.Code())
		if _, err := w.Write([]byte(errstatus.Message())); err != nil {
			w.WriteHeader(http.StatusInternalServerError)
		}

		return
	}

	raw, err := runtime.ProtoMarshal(msg)
	if err != nil {
		w.WriteHeader(http.StatusInternalServerError)
		if _, err := w.Write([]byte(err.Error())); err != nil {
			w.WriteHeader(http.StatusInternalServerError)
		}

		return
	}

	w.WriteHeader(200)
	if _, err := w.Write(raw); err != nil {
		w.WriteHeader(http.StatusInternalServerError)
	}

	return

}

func handlerCallbackServiceWebServerRegisterCallback(server CallbackServiceWebServer, w http.ResponseWriter, r *http.Request, _ runtime.Params, il []runtime.Interceptor) {
	ctx, cancel := context.WithCancel(r.Context())
	defer cancel()

	w.Header().Set("Content-Type", "application/json")

	ctx, err := runtime.ApplyInterceptors(ctx, r, il...)
	if err != nil {
		errstatus := runtime.GetHTTPStatusFromError(err)

		w.WriteHeader(errstatus.Code())
		if _, err := w.Write([]byte(errstatus.Message())); err != nil {
			w.WriteHeader(http.StatusInternalServerError)
		}

		return
	}

	ctx = runtime.ContextWithHeaders(ctx, r.Header)

	var protoReq RegisterCallbackRequest

	defer r.Body.Close()

	data, err := io.ReadAll(r.Body)
	if err != nil {
		w.WriteHeader(http.StatusBadRequest)
		if _, err := w.Write([]byte(err.Error())); err != nil {
			w.WriteHeader(http.StatusInternalServerError)
		}

		return
	}

	if len(data) != 0 {
		fm, err := runtime.GetFieldMaskJS(data)
		if err != nil {
			w.WriteHeader(http.StatusBadRequest)
			if _, err := w.Write([]byte(err.Error())); err != nil {
				w.WriteHeader(http.StatusInternalServerError)
			}

			return
		}

		ctx = runtime.ContextWithFieldMask(ctx, fm)
		err = runtime.ProtoUnmarshal(data, &protoReq)
		if err != nil {
			w.WriteHeader(http.StatusBadRequest)
			if _, err := w.Write([]byte(err.Error())); err != nil {
				w.WriteHeader(http.StatusInternalServerError)
			}

			return
		}
	}

	_, err = server.RegisterCallback(ctx, &protoReq)
	if err != nil {
		errstatus := runtime.GetHTTPStatusFromError(err)

		w.WriteHeader(errstatus.Code())
		if _, err := w.Write([]byte(errstatus.Message())); err != nil {
			w.WriteHeader(http.StatusInternalServerError)
		}

		return
	}

	w.WriteHeader(201)
	if _, err := w.Write(nil); err != nil {
		w.WriteHeader(http.StatusInternalServerError)
	}

	return

}

func handlerCallbackServiceWebServerListCallbacks(server CallbackServiceWebServer, w http.ResponseWriter, r *http.Request, _ runtime.Params, il []runtime.Interceptor) {
	ctx, cancel := context.WithCancel(r.Context())
	defer cancel()

	w.Header().Set("Content-Type", "application/json")

	ctx, err := runtime.ApplyInterceptors(ctx, r, il...)
	if err != nil {
		errstatus := runtime.GetHTTPStatusFromError(err)

		w.WriteHeader(errstatus.Code())
		if _, err := w.Write([]byte(errstatus.Message())); err != nil {
			w.WriteHeader(http.StatusInternalServerError)
		}

		return
	}

	ctx = runtime.ContextWithHeaders(ctx, r.Header)

	var protoReq ListCallbacksRequest

	msg, err := server.ListCallbacks(ctx, &protoReq)
	if err != nil {
		errstatus := runtime.GetHTTPStatusFromError(err)

		w.WriteHeader(errstatus.Code())
		if _, err := w.Write([]byte(errstatus.Message())); err != nil {
			w.WriteHeader(http.StatusInternalServerError)
		}

		return
	}

	raw, err := runtime.ProtoMarshal(msg)
	if err != nil {
		w.WriteHeader(http.StatusInternalServerError)
		if _, err := w.Write([]byte(err.Error())); err != nil {
			w.WriteHeader(http.StatusInternalServerError)
		}

		return
	}

	w.WriteHeader(200)
	if _, err := w.Write(raw); err != nil {
		w.WriteHeader(http.StatusInternalServerError)
	}

	return

}

func handlerCallbackServiceWebServerDeleteCallback(server CallbackServiceWebServer, w http.ResponseWriter, r *http.Request, _ runtime.Params, il []runtime.Interceptor) {
	ctx, cancel := context.WithCancel(r.Context())
	defer cancel()

	w.Header().Set("Content-Type", "application/json")

	ctx, err := runtime.ApplyInterceptors(ctx, r, il...)
	if err != nil {
		errstatus := runtime.GetHTTPStatusFromError(err)

		w.WriteHeader(errstatus.Code())
		if _, err := w.Write([]byte(errstatus.Message())); err != nil {
			w.WriteHeader(http.StatusInternalServerError)
		}

		return
	}

	ctx = runtime.ContextWithHeaders(ctx, r.Header)

	var protoReq DeleteCallbackRequest

	defer r.Body.Close()

	data, err := io.ReadAll(r.Body)
	if err != nil {
		w.WriteHeader(http.StatusBadRequest)
		if _, err := w.Write([]byte(err.Error())); err != nil {
			w.WriteHeader(http.StatusInternalServerError)
		}

		return
	}

	if len(data) != 0 {
		fm, err := runtime.GetFieldMaskJS(data)
		if err != nil {
			w.WriteHeader(http.StatusBadRequest)
			if _, err := w.Write([]byte(err.Error())); err != nil {
				w.WriteHeader(http.StatusInternalServerError)
			}

			return
		}

		ctx = runtime.ContextWithFieldMask(ctx, fm)
		err = runtime.ProtoUnmarshal(data, &protoReq)
		if err != nil {
			w.WriteHeader(http.StatusBadRequest)
			if _, err := w.Write([]byte(err.Error())); err != nil {
				w.WriteHeader(http.StatusInternalServerError)
			}

			return
		}
	}

	_, err = server.DeleteCallback(ctx, &protoReq)
	if err != nil {
		errstatus := runtime.GetHTTPStatusFromError(err)

		w.WriteHeader(errstatus.Code())
		if _, err := w.Write([]byte(errstatus.Message())); err != nil {
			w.WriteHeader(http.StatusInternalServerError)
		}

		return
	}

	w.WriteHeader(204)
	if _, err := w.Write(nil); err != nil {
		w.WriteHeader(http.StatusInternalServerError)
	}

	return

}