# api/rest/gateway_swagger.json
{"components":{"requestBodies":{"DeleteCallbackRequest":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteCallbackRequest"}}},"description":"DeleteCallbackRequest unregisters a webhook URL.","required":true},"RegisterCallbackRequest":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterCallbackRequest"}}},"description":"RegisterCallbackRequest registers a webhook URL.","required":true},"SendMessageRequest":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SendMessageRequest"}}},"description":"SendMessageRequest is a text message to transmit to one node.","required":true}},"responses":{"ListCallbacksResponse":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ListCallbacksResponse"}}},"description":"ListCallbacksResponse carries the registered callbacks."},"ListChannelsResponse.Channels":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ListChannelsResponse.Channels"}}},"description":"Channel is one configured channel on the connected radio."},"ListMessagesResponse":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ListMessagesResponse"}}},"description":"ListMessagesResponse carries the messages for a node."},"ListMessagesResponse.Messages":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ListMessagesResponse.Messages"}}},"description":"Message is one received text message with all stored metadata."},"ListNodesResponse.Nodes":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ListNodesResponse.Nodes"}}},"description":"NodeSummary is the compact node entry returned by the list endpoint."},"Node":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Node"}}},"description":"Node is one full entry in the mesh node database."}},"schemas":{"DeleteCallbackRequest":{"description":"DeleteCallbackRequest unregisters a webhook URL.","properties":{"url":{"description":"webhook URL to remove","type":"string"}},"required":["url"],"type":"object"},"ListCallbacksResponse":{"description":"ListCallbacksResponse carries the registered callbacks.","properties":{"urls":{"items":{"description":"registered webhook subscriptions","type":"string"},"type":"array"}},"required":["urls"],"type":"object"},"ListChannelsResponse.Channels":{"items":{"description":"Channel is one configured channel on the connected radio.","properties":{"index":{"description":"channel index","format":"int32","type":"integer"},"name":{"description":"channel name","type":"string"},"role":{"description":"channel role (e.g. PRIMARY, SECONDARY, DISABLED)","type":"string"}},"required":["index","name","role"],"type":"object"},"type":"array"},"ListMessagesResponse":{"description":"ListMessagesResponse carries the messages for a node.","properties":{"messages":{"items":{"description":"messages, newest first","properties":{"channel":{"description":"channel index the message was received on","format":"int32","type":"integer"},"from":{"description":"sender node number","format":"int32","type":"integer"},"hopsAway":{"description":"hops traversed (hop_start - hop_limit); absent if unknown","format":"int32","type":"integer"},"id":{"description":"on-air Meshtastic packet id (primary key)","format":"int32","type":"integer"},"rssi":{"description":"received signal strength, dBm","format":"int32","type":"integer"},"rxTime":{"description":"receive time, Unix seconds","format":"int64","type":"string"},"snr":{"description":"signal-to-noise ratio of the received packet, dB","type":"number"},"text":{"description":"message body text","type":"string"},"to":{"description":"recipient node number (0xffffffff for broadcast)","format":"int32","type":"integer"}},"required":["id","from","to","channel","text","snr","rssi","rxTime"],"type":"object"},"type":"array"}},"required":["messages"],"type":"object"},"ListMessagesResponse.Messages":{"items":{"description":"Message is one received text message with all stored metadata.","properties":{"channel":{"description":"channel index the message was received on","format":"int32","type":"integer"},"from":{"description":"sender node number","format":"int32","type":"integer"},"hopsAway":{"description":"hops traversed (hop_start - hop_limit); absent if unknown","format":"int32","type":"integer"},"id":{"description":"on-air Meshtastic packet id (primary key)","format":"int32","type":"integer"},"rssi":{"description":"received signal strength, dBm","format":"int32","type":"integer"},"rxTime":{"description":"receive time, Unix seconds","format":"int64","type":"string"},"snr":{"description":"signal-to-noise ratio of the received packet, dB","type":"number"},"text":{"description":"message body text","type":"string"},"to":{"description":"recipient node number (0xffffffff for broadcast)","format":"int32","type":"integer"}},"required":["id","from","to","channel","text","snr","rssi","rxTime"],"type":"object"},"type":"array"},"ListNodesResponse.Nodes":{"items":{"description":"NodeSummary is the compact node entry returned by the list endpoint.","properties":{"longName":{"description":"human-readable long name","type":"string"},"num":{"description":"node number (primary key)","format":"int32","type":"integer"},"shortName":{"description":"short name / call sign","type":"string"}},"required":["num","shortName","longName"],"type":"object"},"type":"array"},"Node":{"description":"Node is one full entry in the mesh node database.","properties":{"altitude":{"description":"altitude, metres; absent if unknown","format":"int32","type":"integer"},"battery":{"description":"battery level, percent; absent if unknown","format":"int32","type":"integer"},"hopsAway":{"description":"hops away from us; absent if unknown","format":"int32","type":"integer"},"hwModel":{"description":"hardware model","type":"string"},"isFavorite":{"description":"true if marked favorite on the radio","type":"boolean"},"lastHeard":{"description":"last time a packet was heard, Unix seconds","format":"int64","type":"string"},"latitude":{"description":"latitude, degrees; absent if unknown","type":"number"},"longName":{"description":"human-readable long name","type":"string"},"longitude":{"description":"longitude, degrees; absent if unknown","type":"number"},"num":{"description":"node number (primary key)","format":"int32","type":"integer"},"role":{"description":"device role","type":"string"},"rssi":{"description":"received signal strength of the last packet, dBm","format":"int32","type":"integer"},"shortName":{"description":"short name / call sign","type":"string"},"snr":{"description":"signal-to-noise ratio of the last packet, dB","type":"number"},"updatedAt":{"description":"last time this record changed, Unix seconds","format":"int64","type":"string"},"viaMqtt":{"description":"true if last seen over MQTT rather than LoRa","type":"boolean"},"voltage":{"description":"battery voltage, volts; absent if unknown","type":"number"}},"required":["num","longName","shortName","hwModel","role","snr","rssi","viaMqtt","isFavorite","lastHeard","updatedAt"],"type":"object"},"RegisterCallbackRequest":{"description":"RegisterCallbackRequest registers a webhook URL.","properties":{"url":{"description":"webhook URL to register","type":"string"}},"required":["url"],"type":"object"},"SendMessageRequest":{"description":"SendMessageRequest is a text message to transmit to one node.","properties":{"channel":{"description":"channel index (0 by default)","format":"int32","type":"integer"},"text":{"description":"message body text","type":"string"}},"required":["text"],"type":"object"}}},"info":{"title":"Swagger API (rest)","version":"20260820"},"openapi":"3.0.3","paths":{"/api/v1/callback":{"delete":{"operationId":"CallbackServiceDeleteCallback","requestBody":{"$ref":"#/components/requestBodies/DeleteCallbackRequest"},"responses":{"204":{},"default":{"description":""}},"summary":"DeleteCallback unsubscribes a webhook URL.","tags":["CallbackService"]},"get":{"operationId":"CallbackServiceListCallbacks","responses":{"200":{"$ref":"#/components/responses/ListCallbacksResponse"},"default":{"description":""}},"summary":"ListCallbacks returns every registered webhook subscriptions.","tags":["CallbackService"]},"post":{"operationId":"CallbackServiceRegisterCallback","requestBody":{"$ref":"#/components/requestBodies/RegisterCallbackRequest"},"responses":{"201":{},"default":{"description":""}},"summary":"RegisterCallback subscribes a webhook URL to mesh events.","tags":["CallbackService"]}},"/api/v1/channels":{"get":{"operationId":"MeshServiceListChannels","responses":{"200":{"$ref":"#/components/responses/ListChannelsResponse.Channels"},"default":{"description":""}},"summary":"ListChannels returns every configured channel on the radio.","tags":["MeshService"]}},"/api/v1/channels/{index}/messages":{"get":{"operationId":"MeshServiceListChannelMessages","parameters":[{"description":"channel index (path)","in":"path","name":"index","required":true,"schema":{"description":"channel index (path)","format":"int32","type":"integer"}}],"responses":{"200":{"$ref":"#/components/responses/ListMessagesResponse"},"default":{"description":""}},"summary":"ListChannelMessages returns messages received on one channel.","tags":["MeshService"]}},"/api/v1/nodes":{"get":{"operationId":"MeshServiceListNodes","responses":{"200":{"$ref":"#/components/responses/ListNodesResponse.Nodes"},"default":{"description":""}},"summary":"ListNodes returns a compact summary of every node the radio can see.","tags":["MeshService"]}},"/api/v1/nodes/{num}":{"get":{"operationId":"MeshServiceGetNode","parameters":[{"description":"node number (path)","in":"path","name":"num","required":true,"schema":{"description":"node number (path)","format":"int32","type":"integer"}}],"responses":{"200":{"$ref":"#/components/responses/Node"},"default":{"description":""}},"summary":"GetNode returns all stored information for one node by its number.","tags":["MeshService"]}},"/api/v1/nodes/{num}/messages":{"get":{"operationId":"MeshServiceListNodeMessages","parameters":[{"description":"sender node number (path)","in":"path","name":"num","required":true,"schema":{"description":"sender node number (path)","format":"int32","type":"integer"}}],"responses":{"200":{"$ref":"#/components/responses/ListMessagesResponse.Messages"},"default":{"description":""}},"summary":"ListNodeMessages returns direct messages received from one node.","tags":["MeshService"]},"post":{"operationId":"MeshServiceSendNodeMessage","parameters":[{"description":"recipient node number (path)","in":"path","name":"num","required":true,"schema":{"description":"recipient node number (path)","format":"int32","type":"integer"}}],"requestBody":{"$ref":"#/components/requestBodies/SendMessageRequest"},"responses":{"202":{},"default":{"description":""}},"summary":"SendNodeMessage transmits a text message to one node.","tags":["MeshService"]}}},"tags":[{"description":"MeshService exposes the connected Meshtastic node.","name":"MeshService"},{"description":"CallbackService manages webhook subscriptions.","name":"CallbackService"}]}